#4: increase floating point precision

+++ adma_ros_driver/include/adma_ros_driver/parser/parser_utils.hpp
@@ -5,4 +5,4 @@
 const float PI=3.1415926535897932384626433832795028841971f;
 
 bool getbit(unsigned char byte, int position);
-float getScaledValue(int32_t rawValue, float lsbFactor);
+double getScaledValue(int32_t rawValue, double lsbFactor);
